+++ Sources/Utils/GUIUtils/GuiEnabledStateScope.cs
@@ -9,13 +9,13 @@ namespace KSPDev.GUIUtils {
 
 /// <summary>A utility class to render big disabled bloacks of GUI.</summary>
 /// <example><code source="Examples/GUIUtils/GuiWindow-Examples.cs" region="GuiEnabledStateDemo1_OnGUI"/></example>
-public class GuiEnabledState : IDisposable {
+public class GuiEnabledStateScope : IDisposable {
   readonly bool oldState;
 
   /// <summary>Stores the old state and sets a new one.</summary>
   /// <param name="newState">The new state to set.</param>
   /// <example><code source="Examples/GUIUtils/GuiWindow-Examples.cs" region="GuiEnabledStateDemo1_OnGUI"/></example>
-  public GuiEnabledState(bool newState) {
+  public GuiEnabledStateScope(bool newState) {
     oldState = GUI.enabled;
     GUI.enabled = newState;
   }